+++ sdk/cosmos/azure-cosmos-encryption/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Release History
 
-### 2.26.0-beta.1 (Unreleased)
+### 2.27.0-beta.1 (Unreleased)
 
 #### Features Added
 
@@ -9,6 +9,11 @@
 #### Bugs Fixed
 
 #### Other Changes
+
+### 2.26.0 (2026-01-26)
+
+#### Other Changes
+* Updated `azure-cosmos` to version `4.77.0`.
 
 ### 2.25.0 (2025-12-09)